+++ changelog.md
@@ -29,6 +29,7 @@
 - Modifies copy in `index.mdx`
 - Creates new page `colophon.mdx` and moves part of `index.mdx` to it
 - Removes `prefetch` directives from site links
+- Adds `<link rel='preload' />` for webfont
 
 ### Known Issues